UserStatus has its own class
