+++ apps/challenge-user-service/.openapi-generator/FILES
@@ -6,6 +6,7 @@ src/main/java/org/sagebionetworks/challenge/api/UserApiDelegate.java
 src/main/java/org/sagebionetworks/challenge/configuration/HomeController.java
 src/main/java/org/sagebionetworks/challenge/configuration/SpringDocConfiguration.java
 src/main/java/org/sagebionetworks/challenge/model/dto/User.java
+src/main/java/org/sagebionetworks/challenge/model/dto/UserStatus.java
 src/main/java/org/sagebionetworks/challenge/model/dto/UserUpdateRequest.java
 src/main/resources/openapi.yaml
 src/test/java/org/sagebionetworks/challenge/OpenApiGeneratorApplicationTests.java
