+++ apps/challenge-user-service/src/main/java/org/sagebionetworks/challenge/model/dto/User.java
@@ -1,12 +1,11 @@
 package org.sagebionetworks.challenge.model.dto;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonValue;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.*;
 import java.util.Objects;
 import javax.annotation.Generated;
+import javax.validation.Valid;
 import javax.validation.constraints.*;
 
 /** User */
@@ -28,45 +27,8 @@ public class User {
   @JsonProperty("authId")
   private String authId;
 
-  /** Gets or Sets status */
-  public enum StatusEnum {
-    PENDING("PENDING"),
-
-    APPROVED("APPROVED"),
-
-    DISABLED("DISABLED"),
-
-    BLACKLIST("BLACKLIST");
-
-    private String value;
-
-    StatusEnum(String value) {
-      this.value = value;
-    }
-
-    @JsonValue
-    public String getValue() {
-      return value;
-    }
-
-    @Override
-    public String toString() {
-      return String.valueOf(value);
-    }
-
-    @JsonCreator
-    public static StatusEnum fromValue(String value) {
-      for (StatusEnum b : StatusEnum.values()) {
-        if (b.value.equals(value)) {
-          return b;
-        }
-      }
-      throw new IllegalArgumentException("Unexpected value '" + value + "'");
-    }
-  }
-
   @JsonProperty("status")
-  private StatusEnum status;
+  private UserStatus status;
 
   public User id(Long id) {
     this.id = id;
@@ -163,7 +125,7 @@ public class User {
     this.authId = authId;
   }
 
-  public User status(StatusEnum status) {
+  public User status(UserStatus status) {
     this.status = status;
     return this;
   }
@@ -173,12 +135,13 @@ public class User {
    *
    * @return status
    */
+  @Valid
   @Schema(name = "status", required = false)
-  public StatusEnum getStatus() {
+  public UserStatus getStatus() {
     return status;
   }
 
-  public void setStatus(StatusEnum status) {
+  public void setStatus(UserStatus status) {
     this.status = status;
   }
